requirements.txt
discord.py==1.2.3
praw==6.4.0
requests==2.22.0
Pillow